Перенести эндпойнт создания аккаунта к эндпойнтам аккаунта (#158)
Во время работы над добавлением сваггера для эндпойнтов авторизации (#77), эндпойнт создания аккаунта был добавлен в пакет `auth`.

Тогда казалось, что этот эндпойнт должен быть там, т.к. он будет использоваться в процессе регистрации. Но если посмотреть на этот эндпойнт более внимательно, то окажется, что с точки зрения REST архитектуры этот эндпойнт не имеет ничего общего с авторизацией и аутентификацией. Этот эндпойнт просто создаёт ресурс и при этом не проверяет и не выдаёт никаких прав пользователю. Этот эндпойнт относится к CRUD набору операций, доступных над сущностью акканута, и, следовательно, должен быть перемещён в пакет `account`.

Это также позволит нам убрать из пакета `auth` данные, которые выглядят инородными для этого пакета, например:
- схемы `NewAccount` и `NewProfile`, которые по логике хотелось бы видеть в пакетах `account` и `profile` соответственно
- поле `Password`, которое по нашей DDD модели должно принадлежать сущности Аккаунта, и следовательно находиться в пакете `account`

В рамках этой задачи необходимо перенести эндпойнт создания аккаунта в пакет с эндпойнтами аккаунта.
