+++ src/account/controllers.py
@@ -4,9 +4,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from src.account import schemas as account_schemas
+from src.account import schemas
 from src.account.models import Account, PasswordHash
-from src.auth import schemas
 from src.profile import schemas as profile_schemas
 from src.profile.models import Profile
 
@@ -14,7 +13,7 @@ from src.profile.models import Profile
 if TYPE_CHECKING:
     from sqlalchemy.ext.asyncio import AsyncSession
 
-    from src.auth.schemas import NewAccountWithProfile
+    from src.account.schemas import NewAccountWithProfile
 
 
 async def create_account(request_data: NewAccountWithProfile, session: AsyncSession) -> schemas.AccountWithProfile:
@@ -23,6 +22,6 @@ async def create_account(request_data: NewAccountWithProfile, session: AsyncSess
     profile = await Profile.create(session, request_data.profile, account_id=account.id)
     await PasswordHash.create(session, request_data.account.password, account_id=account.id)
     return schemas.AccountWithProfile(
-        account=account_schemas.Account.from_orm(account),
+        account=schemas.Account.from_orm(account),
         profile=profile_schemas.Profile.from_orm(profile),
     )
