+++ src/account/routes.py
@@ -6,14 +6,69 @@ from typing import Annotated
 
 from fastapi import APIRouter, Body, Depends, Path, status
 from fastapi.security import HTTPAuthorizationCredentials
+from sqlalchemy.ext.asyncio import AsyncSession
 
-from src.account import schemas
+from src.account import controllers, schemas
 from src.account.fields import Login
 from src.auth.security import get_token
 from src.shared import swagger as shared_swagger
+from src.shared.database import get_session
 
 
 router = APIRouter(tags=["account"])
+
+
+@router.post(
+    "/accounts",
+    description="Sign Up - create a new account with profile.",
+    responses={
+        status.HTTP_201_CREATED: {
+            "description": "New account and profile are created.",
+            "content": {
+                "application/json": {
+                    "example": {
+                        "account": {
+                            "id": 42,
+                            "email": "john.doe@mail.com",
+                            "login": "john_doe",
+                            "created_at": "2023-06-17T11:47:02.823Z",
+                        },
+                        "profile": {
+                            "account_id": 42,
+                            "avatar_id": None,
+                            "description": "Who da heck is John Doe?",
+                            "name": "John Doe",
+                        },
+                    },
+                },
+            },
+        },
+    },
+    response_model=schemas.AccountWithProfile,
+    status_code=status.HTTP_201_CREATED,
+    summary="Sign Up - create a new account.",
+)
+async def create_account(
+    new_account: Annotated[
+        schemas.NewAccountWithProfile,
+        Body(
+            example={
+                "account": {
+                    "email": "john.doe@mail.com",
+                    "login": "john_doe",
+                    "password": "qwerty123",
+                },
+                "profile": {
+                    "name": "John Doe",
+                    "description": "Who da heck is John Doe?",
+                },
+            },
+        ),
+    ],
+    session: AsyncSession = Depends(get_session),
+) -> schemas.AccountWithProfile:
+    """Create a new account with profile."""
+    return await controllers.create_account(new_account, session)
 
 
 @router.get(
